fix: address round-2 PR review feedback
- Filter negated gitignore patterns (!pattern) instead of passing them through
- Replace std::process::exit(1) with anyhow::bail! in cmd_fmt check mode
- Skip empty keys in load_dotenv to prevent set_var panic on "=value" lines
- Skip symlink directories in walk_dir to prevent infinite recursion
- Guard tabWidth/printWidth casts with range checks to prevent silent truncation
- Normalize backslashes to forward slashes in is_ignored for Windows compat
- Extract cmd_fmt tests to cmd_fmt_tests.rs to keep source under 700 lines
